:zap: Update seed #71

+++ db/migrate/20211209050722_create_foods.rb
@@ -3,8 +3,8 @@ class CreateFoods < ActiveRecord::Migration[6.0]
     create_table :foods do |t|
       t.integer :classification
       t.string :food_name
-      t.integer :food_amount
-      t.integer :sugar_amount
+      t.float :food_amount
+      t.float :sugar_amount
 
       t.timestamps
     end
